services/ingestion_api/requirements.txt
fastapi==0.136.0
uvicorn[standard]==0.44.0
pydantic==2.12.5
pydantic-settings==2.13.1
aiokafka==0.13.0
prometheus-client==0.24.1
opentelemetry-api==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
